Utility_Python_Scripts/Poem_to_Image/requirements.txt
google-generativeai
pillow
python-dotenv
